return cache.put promise in chain
after opening the cache, a put is executed and a log line is supposed to be written after the put is complete. `cache.put` returns a promise and returning it will wait the promise chain until that operation is complete and thus log the "response stored" message *after* the put is completed.